Update MAC_SDK from 4.44 to 9.04
This is a larger update of MAC_SDK to release 9.04.
- Original download location:
  https://monkeysaudio.com/files/MAC_904_SDK.zip
  SHA256:
  09cfebe79986a9b8249cff9c3fd79663692e8e23139903df541905a2ecf44c6a
- Several function return values and arguments are 64-bit now in
  MAC_SDK 9.04, e.g. in IO.h:
    int64 GetPosition();
    int64 GetSize();
  Use type `long` accordingly and remove casts, which are not required
  anymore.
- Update the CUETools specific MACLibDll.cpp and MACLibDll.h based on
  the upstream modifications to MACDll.cpp and MACDll.h, respectively.
- Update ThirdParty_MAC_SDK_CUETools.patch accordingly
- CUETools.Codecs.MACLib\AudioEncoder.cs:
  Remove duplicate "m_settings = settings;"
- Resolves #158

+++ CUETools.Codecs.MACLib/AudioDecoder.cs
@@ -89,7 +89,7 @@ namespace CUETools.Codecs.MACLib
                 _bufferOffset = 0;
                 _bufferLength = 0;
                 _sampleOffset = value;
-                int res = MACLibDll.c_APEDecompress_Seek(pAPEDecompress, (int)value);
+                int res = MACLibDll.c_APEDecompress_Seek(pAPEDecompress, value);
                 if (0 != res)
                     throw new Exception("unable to seek:" + res.ToString());
             }
@@ -117,7 +117,7 @@ namespace CUETools.Codecs.MACLib
                     long nBlocksRetrieved;
                     fixed (byte* pSampleBuffer = &_samplesBuffer[0])
                     {
-                        int res = MACLibDll.c_APEDecompress_GetData(pAPEDecompress, (char*)pSampleBuffer, (IntPtr)16384, out nBlocksRetrieved);
+                        int res = MACLibDll.c_APEDecompress_GetData(pAPEDecompress, (char*)pSampleBuffer, 16384, out nBlocksRetrieved);
                         if (res != 0)
                             throw new Exception("An error occurred while decoding: " + res.ToString());
                     }
