+++ CUETools.Codecs.MACLib/StreamIO.cs
@@ -66,14 +66,14 @@ namespace CUETools.Codecs.MACLib
             return 0;
         }
 
-        int TellCallback(void* id)
+        long TellCallback(void* id)
         {
-            return (int)m_stream.Position;
+            return m_stream.Position;
         }
 
-        uint GetSizeCallback(void* id)
+        long GetSizeCallback(void* id)
         {
-            return (uint)m_stream.Length;
+            return m_stream.Length;
         }
 
         long SeekRelativeCallback(void* id, long delta, int mode)
